phase 0b: fix NVML counter type on Linux

+++ tools/ortprobe/ortprobe.cpp
@@ -566,7 +566,8 @@ void sampleVram(VramMonitor *monitor, FlowRun::VramMetrics &metrics, const char 
     metrics.peakUsedBytes = sample.used;
     metrics.hasSample = true;
   } else {
-    metrics.peakUsedBytes = std::max(metrics.peakUsedBytes, sample.used);
+    metrics.peakUsedBytes =
+        std::max(metrics.peakUsedBytes, static_cast<std::uint64_t>(sample.used));
   }
   if (std::strcmp(label, "steady") == 0) {
     metrics.steadyUsedBytes = sample.used;
